+++ src/main/java/baseball/domain/count/Count.java
@@ -1,4 +1,4 @@
-package baseball.domain;
+package baseball.domain.count;
 
 public enum Count {
